refactor(types): added types for better dx
